requirements.txt
pytorch
numpy
glob
sklearn